# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CareMatch API
  version: 0.11.0
  description: |
    REST API for the CareMatch healthcare recruitment platform.

    The first-party SPA authenticates with a Laravel Sanctum stateful
    cookie session. Browser clients must first call `GET /sanctum/csrf-cookie`,
    include credentials on subsequent requests, and send the decoded
    `XSRF-TOKEN` cookie value in the `X-XSRF-TOKEN` header on state-changing
    requests.

    CareMatch does not expose personal access token issuing or management APIs
    in Phase 2-A. Phase 2-B resolves tenant access from the authenticated user,
    the organisation route identifier, and an active persisted membership before
    tenant-owned controllers run. Phase 3A applies that boundary to Candidate
    resources and scopes every nested Candidate identifier to its route Organisation.
    Phase 3B applies the same boundary to Job profiles and their explicit lifecycle.
    Phase 4 adds tenant-safe Applications and the recruitment pipeline. Phase 5A
    adds private Candidate documents. Document contents have no public storage URL;
    every metadata, download, upload, and delete operation is tenant-authorised.
    Phase 5B adds a small tenant-scoped Dashboard read model.

    Deployments configured as the public portfolio demo return `403` for
    registration, Organisation creation, password recovery, invitation creation,
    and Candidate document upload/deletion. These deployment restrictions are
...
  license:
...
servers:
...
paths:
...
  /organisations/{organisation}/qualifications/{qualification}:
    parameters:
      - { $ref: "#/components/parameters/OrganisationId" }
      - { $ref: "#/components/parameters/QualificationId" }
    patch:
      security: [{ StatefulSession: [] }]
      operationId: updateQualificationDefinition
      summary: Update or activate/deactivate a qualification definition
      tags: [Compliance]
      parameters: [{ $ref: "#/components/parameters/XsrfToken" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/QualificationDefinitionUpdateInput" } } } }
      responses:
        "200": { description: Definition updated, content: { application/json: { schema: { $ref: "#/components/schemas/QualificationDefinitionResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/TenantNotFound" }
        "419": { $ref: "#/components/responses/CsrfTokenMismatch" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /organisations/{organisation}/jobs/{job}/qualification-requirements:
    parameters:
      - { $ref: "#/components/parameters/OrganisationId" }
      - { $ref: "#/components/parameters/JobId" }
    get:
      security: [{ StatefulSession: [] }]
      operationId: listJobQualificationRequirements
      summary: List required Job qualifications
      tags: [Compliance]
      responses:
        "200": { description: Required qualifications, content: { application/json: { schema: { $ref: "#/components/schemas/JobQualificationRequirementCollection" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/JobNotFound" }
    post:
      security: [{ StatefulSession: [] }]
      operationId: addJobQualificationRequirement
      summary: Add a required Job qualification
      tags: [Compliance]
      parameters: [{ $ref: "#/components/parameters/XsrfToken" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/JobQualificationRequirementInput" } } } }
      responses:
        "201": { description: Requirement added, content: { application/json: { schema: { $ref: "#/components/schemas/JobQualificationRequirementResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/JobNotFound" }
        "409": { description: The qualification is already required, content: { application/json: { schema: { $ref: "#/components/schemas/SimpleMessage" } } } }
        "419": { $ref: "#/components/responses/CsrfTokenMismatch" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    StatefulSession:
      type: apiKey
      in: cookie
      name: carematch_session
      description: |
        Laravel Sanctum stateful authentication for the first-party SPA. The
        HttpOnly session cookie is issued and rotated by the server. Browser
        clients must send requests with credentials. This scheme does not
        represent a Bearer token or a personal access token.
  parameters:
    OrganisationId:
      name: organisation
      in: path
      required: true
      description: The organisation identifier resolved against the authenticated user's active membership.
      schema:
        type: integer
        format: int64
        minimum: 1
...
    JobId:
      name: job
      in: path
      required: true
      description: Job identifier resolved only within the trusted route Organisation.
      schema:
        type: integer
        format: int64
        minimum: 1
...
    QualificationId:
      name: qualification
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: |
        URL-decoded value of the `XSRF-TOKEN` cookie obtained from
        `GET /sanctum/csrf-cookie`. Required by Laravel CSRF protection for
        state-changing first-party SPA requests.
      schema:
        type: string
        minLength: 1
...
  responses:
...
    ValidationFailed:
      description: The submitted input failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Unauthenticated:
      description: No authenticated stateful session is available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthenticatedError"
    Forbidden:
      description: |
        The active tenant member does not have the required role, or this write
        operation is disabled by the public portfolio demo deployment policy.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenError"
    TenantNotFound:
      description: |
        The organisation is nonexistent or is not available through an active
        membership for the authenticated user. These cases are intentionally
        indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
...
    JobNotFound:
      description: |
        The Job does not exist within the trusted route Organisation. Cross-tenant
        and nonexistent identifiers are intentionally indistinguishable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/NotFoundError" }
...
    CsrfTokenMismatch:
      description: The CSRF token is missing, invalid, or no longer matches the session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CsrfError"
...
  schemas:
    SimpleMessage:
      type: object
      required: [message]
      properties: { message: { type: string } }
    QualificationDefinition:
      type: object
      additionalProperties: false
      required: [id, name, category, description, is_active, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 200 }
        category: { type: [string, "null"], maxLength: 100 }
        description: { type: [string, "null"], maxLength: 1000 }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    QualificationDefinitionInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        category: { type: [string, "null"], maxLength: 100 }
        description: { type: [string, "null"], maxLength: 1000 }
        is_active: { type: boolean, default: true }
    QualificationDefinitionResponse:
      type: object
      required: [data]
      properties: { data: { $ref: "#/components/schemas/QualificationDefinition" } }
    QualificationDefinitionUpdateInput:
      allOf:
        - $ref: "#/components/schemas/QualificationDefinitionInput"
        - type: object
          required: [name, is_active]
...
    JobQualificationRequirementInput:
      type: object
      additionalProperties: false
      required: [qualification_definition_id]
      properties: { qualification_definition_id: { type: integer, format: int64, minimum: 1 } }
    JobQualificationRequirement:
      type: object
      required: [id, qualification_definition_id, qualification_name, created_at]
      properties:
        id: { type: integer, format: int64 }
        qualification_definition_id: { type: integer, format: int64 }
        qualification_name: { type: string }
        created_at: { type: string, format: date-time }
    JobQualificationRequirementResponse:
      type: object
      required: [data]
      properties: { data: { $ref: "#/components/schemas/JobQualificationRequirement" } }
    JobQualificationRequirementCollection:
      type: object
      required: [data]
      properties: { data: { type: array, items: { $ref: "#/components/schemas/JobQualificationRequirement" } } }
...
    ValidationError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          description: Validation messages keyed by request field.
          additionalProperties:
            type: array
            items:
              type: string
    UnauthenticatedError:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          const: Unauthenticated.
    ForbiddenError:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          examples:
            - This action is unauthorized.
            - Registration is unavailable in the public portfolio demo.
    NotFoundError:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          const: Not Found
...
    CsrfError:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          const: CSRF token mismatch.
...